Ćwiczenie 34.3

+++ src/products/products.controller.ts
@@ -1,6 +1,17 @@
-import { Controller, Get, Param, Delete, Post, Body } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Param,
+  Delete,
+  Post,
+  Body,
+  Put,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from 'src/dtos/create-product.dto';
+import { UpdateProductDTO } from 'src/dtos/update-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
+import { NotFoundException } from '@nestjs/common';
 
 @Controller('products')
 export class ProductsController {
@@ -8,19 +19,27 @@ export class ProductsController {
 
   // GET(/products) - get all products
   @Get('/')
-  getAll(): any {
+  getAllProducts(): any {
     return this.productsService.getAll();
   }
 
   // GET(/products/:productId) - get product by id
   @Get('/:id')
-  getById(@Param('id') id: string) {
+  getProductById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found!');
+    }
+
     return this.productsService.getById(id);
   }
 
   // DELETE(/products/:productId) - delete product by id
   @Delete('/:id')
-  deleteById(@Param('id') id: string) {
+  deleteProductById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found!');
+    }
+
     this.productsService.deleteById(id);
     return { success: true };
   }
@@ -29,5 +48,19 @@ export class ProductsController {
   @Post('/')
   postNewProduct(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+
+  // PUT(/products/:productId) - edit product by id
+  @Put('/:id')
+  editExistingProduct(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found!');
+    }
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
   }
 }
